ID-1: add Events PresenterService
