+++ app/services/events/presenter_service.rb
@@ -1,0 +1,7 @@
+module Events
+  class PresenterService
+    def self.events
+      Event.on_current_week.to_a
+    end
+  end
+end
